third_party/WebKit/LayoutTests/http/tests/inspector/console-switch-contexts.html

<html>
<head>
<script src="console-test.js"></script>
<script src="inspector-test.js"></script>
<script>
function removeFrame() {
    var frame = document.querySelector("#myFrame");
    frame.parentElement.removeChild(frame);
}

function test()
{
    InspectorTest.showConsolePanel();

    var consoleView = WebInspector.ConsolePanel._view();
    var selector = consoleView._executionContextSelector._selectElement;
    var contextList = WebInspector.runtimeModel.contextLists()[1];
    var contextId = 5;

    function checkAndDumpExecutionContextSelector()
    {
        var option = selector.firstChild;
        while (option) {
            InspectorTest.addResult(option.textContent);
            option = option.nextSibling;
        }
        var selectedOption = consoleView._executionContextSelector.selectedOption();
        InspectorTest.addResult("Selected option: " +  (selectedOption ? selectedOption.textContent : "none!"));
        InspectorTest.assertEquals(WebInspector.runtimeModel.currentExecutionContext(), consoleView._currentExecutionContext());
    }

    function addFakeExecutionContext()
    {
        contextId++;
        var executionContext = new WebInspector.ExecutionContext(WebInspector.targetManager.activeTarget(), contextId, "fake-execution-context " + contextId , false);
        contextList._addExecutionContext(executionContext);
    }

    InspectorTest.addResult("Initial State:")
    checkAndDumpExecutionContextSelector();

    InspectorTest.addResult("Test execution context addition:")
    addFakeExecutionContext();
    addFakeExecutionContext();

    consoleView._executionContextSelector.select(selector.children[2]);
    consoleView._executionContextChanged();

    addFakeExecutionContext();
    addFakeExecutionContext();

    checkAndDumpExecutionContextSelector();

    InspectorTest.addResult("Test frame navigation:")

    contextList._frame._name = "changed frame " + contextList._frame._name;
    contextList._frameNavigated(contextList._frame);

    checkAndDumpExecutionContextSelector();

    InspectorTest.addResult("Test frame detach: ")
    InspectorTest.evaluateInPage("removeFrame()", onFrameRemoved);

    function onFrameRemoved() {
        checkAndDumpExecutionContextSelector();
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body>
<p>

</p>
<iframe name="myIFrame" id="myFrame" src="http://localhost:8000/inspector/resources/console-cd-completions-iframe.html" onload="runTest()"></iframe>

</body>
</html>
