third_party/WebKit/LayoutTests/http/tests/inspector/console-switch-contexts-expected.txt

 Initial State:
<top frame>
myIFrame( console-cd-completions-iframe.html )
Selected option: <top frame>
Test execution context addition:
<top frame>
myIFrame( console-cd-completions-iframe.html )
    fake-execution-context 6
    fake-execution-context 7
    fake-execution-context 8
    fake-execution-context 9
Selected option:     fake-execution-context 6
Test frame navigation:
<top frame>
changed frame myIFrame( c…ompletions-iframe.html )
Selected option: changed frame myIFrame( c…ompletions-iframe.html )
Test frame detach: 
<top frame>
Selected option: <top frame>
